+++ CHANGELOG.md
@@ -23,10 +23,12 @@ versioned in accordance with [Semantic Versioning](https://semver.org/spec/v2.0.
 - BaseConfiguration now provides a `json_encoders` static method for easily accessing the default encoders.
 - The Vegeta connector now validates the `target` and `targets` settings.
 - The check command now supports a verbose and non-verbose modes for outputting multiple checks from a connector.
+- The version command can now output version data for connectors.
 
 ### Changed
 - The reporting interval is now configurable `VegetaConfiguration`.
 - The check event now returns a list of checks instead of one result.
+- Removed the 'Connector` suffix from the default connector name attribute and introduced `full_name`.
 
 ### Removed
 - The `durationpy` package has been removed in favor of a local implementation.